Fix MSVC C4244 narrowing in test_smelt_credit_ignores_claim_winner_identity
The credit-split test I added assigned int worker → int8_t last_towed_by
/ last_fractured_by without a cast, tripping MSVC /WX on two lines.
Other tests in the file that assign these fields all use integer
literals that fit in int8_t so don't warn.

Cast both assignments to (int8_t) explicitly — MSVC stops flagging,
no behavioural change on other compilers.

Windows was the only remaining red job on run 24789006715 after the
broader MSVC fix in ec4350a. 286/286 tests still pass.

+++ src/test_main.c
@@ -908,9 +908,9 @@ TEST(test_smelt_credit_ignores_claim_winner_identity) {
     sha256_bytes(w->players[bystander].session_token, 8, a->fragment_pub);
     /* Worker did both fracture and tow. */
     memcpy(a->last_towed_token, w->players[worker].session_token, 8);
-    a->last_towed_by = worker;
+    a->last_towed_by = (int8_t)worker;
     memcpy(a->last_fractured_token, w->players[worker].session_token, 8);
-    a->last_fractured_by = worker;
+    a->last_fractured_by = (int8_t)worker;
 
     initial_bystander = ledger_balance(&w->stations[0],
                                        w->players[bystander].session_token);
